refactor: actions 코드 분류 및 관련 컴포넌트 import 구문 수정
