+++ src/actions/gatherings/fetchMyPageInfo.ts
