+++ src/actions/gatherings/quitMeeting.ts
