+++ src/app/(myPage)/mypage/page.tsx
@@ -1,6 +1,7 @@
-import { fetchMyPageInfo } from "@/actions/fetchMyPageInfo"
+import { fetchMyPageInfo } from "@/actions/Gatherings/fetchMyPageInfo"
 import MyPageInfoTap from "@/components/pages/mypage/MyPageInfoTap"
 import ProfileBox from "@/components/public/ProfileBox"
+import LIMIT from "@/constants/limit"
 import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query"
 
 const MyPage = async () => {
@@ -8,7 +9,7 @@ const MyPage = async () => {
   await queryClient.prefetchQuery({
     queryKey: ["mypage", {}],
     queryFn: () => {
-      return fetchMyPageInfo({ offset: 0, limit: 5 })
+      return fetchMyPageInfo({ offset: 0, limit: LIMIT })
     },
   })
